.github/actions/process-templates-action/requirements.txt
pyyaml
typer
google-cloud
google-cloud-storage